wdio-cucumber-framework: move out utility functions from reporter class
