+++ packages/wdio-cucumber-framework/src/reporter.js
@@ -1,8 +1,7 @@
 import { Status } from 'cucumber'
 
 import CucumberEventListener from './cucumberEventListener'
-import { createStepArgument } from './utils'
-import * as path from 'path'
+import { createStepArgument, getTestParent, getTestStepTitle, getUniqueIdentifier, formatMessage } from './utils'
 
 class CucumberReporter {
     gherkinDocEvents = []
@@ -29,7 +28,7 @@ class CucumberReporter {
         this.featureStart = new Date()
 
         this.emit('suite:start', {
-            uid: this.getUniqueIdentifier(feature),
+            uid: getUniqueIdentifier(feature),
             title: this.getTitle(feature),
             type: 'suite',
             file: uri,
@@ -44,9 +43,9 @@ class CucumberReporter {
         this.testStart = new Date()
 
         this.emit('suite:start', {
-            uid: this.getUniqueIdentifier(scenario),
+            uid: getUniqueIdentifier(scenario),
             title: this.getTitle(scenario),
-            parent: this.getUniqueIdentifier(feature),
+            parent: getUniqueIdentifier(feature),
             type: 'suite',
             file: uri,
             tags: scenario.tags
@@ -57,11 +56,11 @@ class CucumberReporter {
         this.testStart = new Date()
 
         this.emit('test:start', {
-            uid: this.getUniqueIdentifier(step),
-            title: this.getTestStepTitle(step),
+            uid: getUniqueIdentifier(step),
+            title: getTestStepTitle(step),
             type: 'test',
             file: uri,
-            parent: this.getTestParent(feature, scenario),
+            parent: getTestParent(feature, scenario),
             tags: scenario.tags,
             featureName: feature.name,
             scenarioName: scenario.name,
@@ -85,7 +84,7 @@ class CucumberReporter {
             e = 'pending'
         }
         let error = {}
-        let stepTitle = this.getTestStepTitle(step)
+        let stepTitle = getTestStepTitle(step)
 
         /**
          * if step name is undefined we are dealing with a hook
@@ -139,9 +138,9 @@ class CucumberReporter {
             }
         }
 
-        const parent = this.getTestParent(feature, scenario)
+        const parent = getTestParent(feature, scenario)
         const payload = {
-            uid: this.getUniqueIdentifier(step),
+            uid: getUniqueIdentifier(step),
             title: stepTitle,
             type: 'test',
             file: uri,
@@ -158,9 +157,9 @@ class CucumberReporter {
 
     handleAfterScenario (uri, feature, scenario, sourceLocation) {
         this.emit('suite:end', {
-            uid: this.getUniqueIdentifier(scenario, sourceLocation),
+            uid: getUniqueIdentifier(scenario, sourceLocation),
             title: this.getTitle(scenario),
-            parent: this.getUniqueIdentifier(feature),
+            parent: getUniqueIdentifier(feature),
             type: 'suite',
             file: uri,
             duration: new Date() - this.scenarioStart,
@@ -170,7 +169,7 @@ class CucumberReporter {
 
     handleAfterFeature (uri, feature) {
         this.emit('suite:end', {
-            uid: this.getUniqueIdentifier(feature),
+            uid: getUniqueIdentifier(feature),
             title: this.getTitle(feature),
             type: 'suite',
             file: uri,
@@ -180,7 +179,7 @@ class CucumberReporter {
     }
 
     emit (event, payload) {
-        let message = this.formatMessage({ type: event, payload })
+        let message = formatMessage({ type: event, payload })
 
         message.cid = this.cid
         message.specs = this.specs
@@ -194,94 +193,6 @@ class CucumberReporter {
         const tags = featureOrScenario.tags
         if (!this.tagsInTitle || !tags.length) return name
         return `${tags.map(tag => tag.name).join(', ')}: ${name}`
-    }
-
-    getUriOf (type) {
-        if (!type || !type.uri) {
-            return
-        }
-
-        return type.uri.replace(process.cwd(), '')
-    }
-
-    getUniqueIdentifier (target, sourceLocation) {
-        if (target.type === 'Hook') {
-            return `${path.basename(target.location.uri)}${target.location.line}`
-        }
-
-        if (target.type === 'ScenarioOutline') {
-            let name = target.name || target.text
-            const line = sourceLocation.line || ''
-
-            if (Array.isArray(target.examples)) {
-                target.examples[0].tableHeader.cells.forEach((header, idx) => {
-                    if (name.indexOf('<' + header.value + '>') === -1) {
-                        return
-                    }
-
-                    target.examples[0].tableBody.forEach((tableEntry) => {
-                        if (tableEntry.location.line === sourceLocation.line) {
-                            name = name.replace('<' + header.value + '>', tableEntry.cells[idx].value)
-                        }
-                    })
-                })
-            }
-
-            return `${name}${line}`
-        }
-
-        const name = target.name || target.text
-        const location = target.location || target.locations[0]
-        const line = (location && location.line) || ''
-        return `${name}${line}`
-    }
-
-    getTestParent(feature, scenario) {
-        return `${feature.name || 'Undefined Feature'} > ${scenario.name || 'Undefined Scenario'}`
-    }
-
-    getTestStepTitle(step) {
-        return ((step.keyword || '') + (step.text || 'Undefined Step')).trim()
-    }
-
-    getTestFullTitle(parent, stepTitle) {
-        return `${parent} > ${stepTitle}`
-    }
-
-    formatMessage ({ type, payload = {} }) {
-        let message = {
-            ...payload,
-            type: type
-        }
-
-        if (payload.error) {
-            message.error = payload.error
-
-            /**
-             * hook failures are emitted as "test:fail"
-             */
-            if (payload.title && payload.title.match(/^"(before|after)( all)*" hook/g)) {
-                message.type = 'hook:end'
-            }
-        }
-
-        /**
-         * Add the current test title to the payload for cases where it helps to
-         * identify the test, e.g. when running inside a beforeEach hook
-         */
-        if (payload.ctx && payload.ctx.currentTest) {
-            message.currentTest = payload.ctx.currentTest.title
-        }
-
-        if (type.match(/Test/)) {
-            message.passed = (payload.state === 'passed')
-        }
-
-        if (payload.title && payload.parent) {
-            payload.fullTitle = this.getTestFullTitle(payload.parent, payload.title)
-        }
-
-        return message
     }
 }
