+++ packages/wdio-cucumber-framework/tests/utils.test.js
@@ -1,4 +1,4 @@
-import { createStepArgument, compareScenarioLineWithSourceLine, getStepFromFeature } from '../src/utils'
+import { createStepArgument, compareScenarioLineWithSourceLine, getStepFromFeature, getTestParent, formatMessage, getUniqueIdentifier, getUriOf } from '../src/utils'
 
 describe('utils', () => {
     describe('createStepArgument', () => {
@@ -59,6 +59,90 @@ describe('utils', () => {
     describe('getStepFromFeature', () => {
         it('should be ok if targetStep.type is not Step', () => {
             getStepFromFeature({ children: [{ type: 'Foo', steps: [{ type: 'Bar' }] }] }, null, 0)
+        })
+    })
+
+    describe('getTestParent', () => {
+        it('no feature and not scenario', () => {
+            expect(getTestParent({}, {})).toEqual('Undefined Feature: Undefined Scenario')
+        })
+    })
+
+    describe('formatMessage', () => {
+        it('should set title for custom type', () => {
+            expect(formatMessage({
+                type: 'foobar',
+                payload: { ctx: { currentTest: { title: 'barfoo' } } }
+            })).toMatchSnapshot()
+        })
+
+        it('should set type to hook:end', () => {
+            expect(formatMessage({
+                type: 'foobar',
+                payload: {
+                    title: '"before all" hook',
+                    error: new Error('boom'),
+                    state: 'passed'
+                }
+            }).type).toBe('hook:end')
+        })
+
+        it('should set passed state for test hooks', () => {
+            expect(formatMessage({
+                type: 'afterTest',
+                payload: { state: 'passed' }
+            })).toMatchSnapshot()
+        })
+
+        it('should not fail if payload was not passed', () => {
+            expect(formatMessage({ type: 'test' })).toMatchSnapshot()
+        })
+    })
+
+    it('getUriOf', () => {
+        expect(getUriOf()).toBe(undefined)
+        expect(getUriOf({})).toBe(undefined)
+        expect(getUriOf({ uri: __filename }))
+            .toBe('/packages/wdio-cucumber-framework/tests/utils.test.js')
+    })
+
+    describe('getUniqueIdentifier', () => {
+        it('Hook', () => {
+            expect(getUniqueIdentifier({
+                type: 'Hook',
+                location: {
+                    uri: __filename,
+                    line: 54
+                }
+            })).toBe('utils.test.js54')
+        })
+
+        it('ScenarioOutline with text', () => {
+            expect(getUniqueIdentifier({
+                type: 'ScenarioOutline',
+                text: 'no-name'
+            }, {})).toBe('no-name')
+        })
+
+        it('ScenarioOutline with <>', () => {
+            expect(getUniqueIdentifier({
+                type: 'ScenarioOutline',
+                name: '<someval> here',
+                examples: [{
+                    tableHeader: {
+                        cells: [
+                            { value: 'valsome' },
+                            { value: 'someval' }
+                        ]
+                    },
+                    tableBody: [{
+                        location: { line: 54 }
+                    }, {
+                        location: { line: 123 },
+                        cells: [{}, { value: 'realval' }]
+                    }]
+                }]
+            }, { line: 123 })).toBe('realval here123')
         })
     })
 })
